+++ src/PythonMigrationViewExtension/PythonMigrationViewExtension.csproj
@@ -193,9 +193,9 @@
   <Target Name="AfterBuild">
     <ItemGroup>
       <ExtensionDefinition Include="PythonMigration_ViewExtensionDefinition.xml" />
-      <CustomFixers Include="$(SolutionDir)PythonMigrationViewExtension\MigrationAssistant\custom_fixers\*" />
+      <PythonFixers Include="$(SolutionDir)PythonMigrationViewExtension\MigrationAssistant\python_migration_fixers\*" />
     </ItemGroup>
     <Copy SourceFiles="@(ExtensionDefinition)" DestinationFolder="$(OutputPath)\viewExtensions\" />
-    <Copy SourceFiles="@(CustomFixers)" DestinationFolder="$(OutputPath)custom_fixers\" />
+    <Copy SourceFiles="@(PythonFixers)" DestinationFolder="$(OutputPath)python_migration_fixers\" />
   </Target>
 </Project>
